Updating _clean to actually clean (#91)
I am not sure what's going on here, but it looks like it wasn't ever cleaning the text. I was trying to embed stuff with

```
featurizer, models = model_from_directory(paper_embedder_dir, on_cpu=True)
embedder = EmbeddingModel(featurizer, models['embedding'])
```

And I noticed that upper case text was not being embedded. Then I chased this down and saw that it was never cleaning the text and never lower-casing it! Any idea what this was for?